Update dependencies as a result of breaking changes in go1.15
The caddyserver packages 'caddy' and 'certmagic' are not compatible
with each other in the latest versions. certmagic@0.8.0 is the latest
working version, so it is pinned. (There also seems to be a name change
from mholt to caddyserver which was not correctly applied everywhere in
the author's code, which is also reflected in the pin.)